# Parametrize JuMP model in optimizer type