The optimizer type allows zero overhead on the JuMP side.
The variable type therefore now need to be parametrized on the model type so that its field are concretely typed which induces many changes of this PR.

### Benchmarking results
This use the benchmarking file in `test/perf/backend_overhead.jl`.
Before this change:
```
v0.6
14.526 ns (0 allocations: 0 bytes)
43.865 ns (5 allocations: 96 bytes)
v0.7
19.497 ns (0 allocations: 0 bytes)
35.113 ns (3 allocations: 64 bytes)
```
After this change:
```
v0.6
1.832 ns (0 allocations: 0 bytes)
24.567 ns (3 allocations: 64 bytes)
v0.7
1.824 ns (0 allocations: 0 bytes)
11.284 ns (1 allocation: 32 bytes)
```

Related to https://github.com/JuliaOpt/MathOptInterface.jl/issues/321