@blegat 2018-11-13
The overhead of having an untyped backend is now 1/3 for both time and space in `@variable(model)`.
With Julia v1.0.2:
```julia
julia> using JuMP

julia> const model = Model()
A JuMP Model

julia> using BenchmarkTools

julia> @btime @variable(model)
  33.454 ns (2 allocations: 48 bytes)
noname
```
By annotating the type of the backend in https://github.com/JuliaOpt/JuMP.jl/blob/master/src/variables.jl#L206:
```julia
julia> using JuMP

julia> const model = Model()
A JuMP Model

julia> using BenchmarkTools

julia> @btime @variable(model)
  18.558 ns (1 allocation: 32 bytes)
noname
```