@mlubin 2018-11-13
I was looking into this a bit last night. I couldn't figure out why exactly Julia is allocating (I couldn't make a small example of an allocation when dispatching on an object whose type is unknown). Maybe there's a trick to fix it.

Another option without parameterizing the model is to do the work in batch and call `add_variables` instead of `add_variable` when we want to create 1M variables.