@blegat 2018-11-13
The batch idea could help. For the 32 remaining bytes, it is due to the fact that VariableRef is not `isbits` because it holds the models as a field and the model is not isbits