@odow 2021-09-27
Closing as stale. There are open PRs exploring alternatives: #2520 and #2610